+++ plone/app/event/tests/test_recurrence.py
@@ -179,14 +179,14 @@ class TestOccurrences(unittest.TestCase):
 
         view = zope.component.getMultiAdapter(
             (self.portal['interval'], self.request), name='event_view')
-        result = view.occurrences
+        result = view.next_occurrences
         # altogether 5 occurrences, but start occurrence is not included
         self.assertEqual(4, len(result['events']))
         self.assertFalse(result['events'][-1] == result['tail'])
 
         view = zope.component.getMultiAdapter(
             (self.portal['many'], self.request), name='event_view')
-        result = view.occurrences
+        result = view.next_occurrences
         self.assertEqual(7, len(result['events']))
         self.assertFalse(result['events'][-1] == result['tail'])